+++ mjx/mujoco/mjx/__init__.py
@@ -30,9 +30,12 @@ from mujoco.mjx._src.forward import step
 from mujoco.mjx._src.inverse import inverse
 from mujoco.mjx._src.io import get_data
 from mujoco.mjx._src.io import get_data_into
+from mujoco.mjx._src.io import get_state
 from mujoco.mjx._src.io import make_data
 from mujoco.mjx._src.io import put_data
 from mujoco.mjx._src.io import put_model
+from mujoco.mjx._src.io import set_state
+from mujoco.mjx._src.io import state_size
 from mujoco.mjx._src.passive import passive
 from mujoco.mjx._src.ray import ray
 from mujoco.mjx._src.sensor import sensor_acc