+++ src/xml/xml_native_reader.cc
@@ -859,7 +859,8 @@ const mjMap fcomp_map[mjNFCOMPTYPES] = {
 const mjMap fdof_map[mjNFCOMPDOFS] = {
   {"full",          mjFCOMPDOF_FULL},
   {"radial",        mjFCOMPDOF_RADIAL},
-  {"trilinear",     mjFCOMPDOF_TRILINEAR}
+  {"trilinear",     mjFCOMPDOF_TRILINEAR},
+  {"quadratic",     mjFCOMPDOF_QUADRATIC}
 };
 
 
@@ -2722,6 +2723,9 @@ void mjXReader::OneFlexcomp(XMLElement* elem, mjsBody* body, const mjVFS* vfs) {
     ReadAttr(elasticity, "damping", 1, &dflex.damping, text);
     ReadAttr(elasticity, "thickness", 1, &dflex.thickness, text);
     MapValue(elasticity, "elastic2d", &dflex.elastic2d, elastic2d_map, 4);
+    if (fcomp.doftype == mjFCOMPDOF_QUADRATIC) {
+      throw mjXError(elasticity, "elasticity is not yet supported for quadratic flex");
+    }
   }
 
   // check errors